pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-receptionist"
version = "0.1.0"
description = "High-fidelity AI phone receptionist using OpenAI Realtime API and LiveKit"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.5.0",
    "livekit-plugins-openai>=1.5.0",
    "livekit-plugins-noise-cancellation>=0.2.3",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "aiosmtplib>=3.0",
    "resend>=2.0",
    "httpx>=0.27",
    "aioboto3>=13.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.12",
    "respx>=0.21",
    "moto>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["receptionist"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
